+++ src/newtab/store/newtab.ts
@@ -1,6 +1,17 @@
 import { IState, reducer } from 'reducers/newtab'
 import { createStore } from 'redux'
+import { loadState, saveState } from 'store/localStorage'
 
-const store = createStore<IState, any, any, any>(reducer)
+const persistedState = loadState();
+const store = createStore<IState, any, any, any>(
+    reducer,
+    persistedState
+);
+
+store.subscribe(() => {
+    saveState({
+        grid: store.getState().grid,
+    });
+});
 
 export default store